# API/app/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CoffeeFlow API Docs",
    "description": "Documentacion interactiva completa de la API de CoffeeFlow: Autenticacion, Usuarios, Mesas, Productos, Pedidos, Cocina (ingredientes/recetas/inventario/alertas), Caja (turnos/tickets/pagos/cortes/gastos/compras) y Notificaciones.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Servidor Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/caja": {
      "get": {
        "tags": [
          "Caja - Turnos y Tickets"
        ],
        "summary": "Listar turnos de caja",
        "description": "Administrador o cajero.",
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "abierto",
                "cerrado"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado de turnos"
          }
        }
      },
      "post": {
        "tags": [
          "Caja - Turnos y Tickets"
        ],
        "summary": "Abrir turno de caja",
        "description": "Administrador o cajero.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fondo_inicial": {
                    "type": "number",
                    "example": 500
                  },
                  "observaciones": {
                    "type": "string",
                    "example": "Turno matutino"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Turno abierto"
          }
        }
      }
    },
    "/api/caja/{id_caja}": {
      "get": {
        "tags": [
          "Caja - Turnos y Tickets"
        ],
        "summary": "Obtener turno de caja",
        "description": "Administrador o cajero.",
        "parameters": [
          {
            "name": "id_caja",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Turno"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
    "/api/caja/{id_caja}/cerrar": {
      "patch": {
        "tags": [
          "Caja - Turnos y Tickets"
        ],
        "summary": "Cerrar turno de caja",
        "description": "Administrador o cajero.",
        "parameters": [
          {
            "name": "id_caja",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Turno cerrado"
          },
          "409": {
            "description": "Ya estaba cerrado"
          }
        }
      }
    },
    "/api/tickets": {
      "get": {
        "tags": [
          "Caja - Turnos y Tickets"
        ],
        "summary": "Listar tickets",
        "description": "Administrador o cajero.",
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pendiente",
                "pagado"
              ]
            }
          },
          {
            "name": "id_caja",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado de tickets"
          }
        }
      },
      "post": {
        "tags": [
          "Caja - Turnos y Tickets"
        ],
        "summary": "Crear ticket",
        "description": "Administrador o cajero. Requiere una caja en estado abierto.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_caja": {
                    "type": "integer",
                    "example": 7
                  },
                  "id_pedido": {
                    "type": "integer",
                    "example": 132
                  },
                  "total": {
                    "type": "number",
                    "example": 197.2
                  },
                  "impuesto": {
                    "type": "number",
                    "example": 27.2
                  },
                  "descuento": {
                    "type": "number",
                    "example": 0
                  }
                },
                "required": [
                  "id_caja",
                  "total"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket creado"
          },
          "400": {
            "description": "Caja no existe o no esta abierta"
          }
        }
      }
    },
    "/api/tickets/{id_ticket}": {
      "get": {
        "tags": [
          "Caja - Turnos y Tickets"
        ],
        "summary": "Obtener ticket (con pagos)",
        "description": "Administrador o cajero.",
        "parameters": [
          {
            "name": "id_ticket",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket con lista de pagos"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
    "/api/pagos": {
      "post": {
        "tags": [
          "Caja - Turnos y Tickets"
        ],
        "summary": "Registrar pago de ticket",
        "description": "Administrador o cajero. Cuando la suma de pagos alcanza el total, el ticket pasa a 'pagado'.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_ticket": {
                    "type": "integer",
                    "example": 110
                  },
                  "monto": {
                    "type": "number",
                    "example": 197.2
                  },
                  "tipo_pago": {
                    "type": "string",
                    "example": "efectivo"
                  },
                  "referencia": {
                    "type": "string",
                    "nullable": true
                  },
                  "cambio": {
                    "type": "number",
                    "example": 2.8
                  }
                },
                "required": [
                  "id_ticket",
                  "monto",
                  "tipo_pago"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pago registrado"
          },
          "404": {
            "description": "Ticket no encontrado"
          },
          "409": {
            "description": "El ticket ya fue pagado"
          }
        }
      }
    },
    "/api/cortes-caja": {
      "get": {
        "tags": [
          "Caja - Cortes, Gastos y Compras"
        ],
        "summary": "Listar cortes de caja",
        "description": "Administrador o cajero.",
        "responses": {
          "200": {
            "description": "Listado de cortes"
          }
        }
      },
      "post": {
        "tags": [
          "Caja - Cortes, Gastos y Compras"
        ],
        "summary": "Crear corte de caja",
        "description": "Administrador o cajero. Calcula automaticamente los totales por metodo de pago a partir de los tickets pagados de la caja.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_caja": {
                    "type": "integer",
                    "example": 7
                  },
                  "diferencia": {
                    "type": "number",
                    "example": 0
                  }
                },
                "required": [
                  "id_caja"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Corte generado"
          },
          "404": {
            "description": "Caja no encontrada"
          }
        }
      }
    },
    "/api/gastos": {
      "get": {
        "tags": [
          "Caja - Cortes, Gastos y Compras"
        ],
        "summary": "Listar gastos",
        "description": "Administrador o cajero.",
        "responses": {
          "200": {
            "description": "Listado de gastos"
          }
        }
      },
      "post": {
        "tags": [
          "Caja - Cortes, Gastos y Compras"
        ],
        "summary": "Registrar gasto",
        "description": "Administrador o cajero.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_caja": {
                    "type": "integer",
                    "example": 7
                  },
                  "concepto": {
                    "type": "string",
                    "example": "Compra de servilletas y vasos"
                  },
                  "monto": {
                    "type": "number",
                    "example": 350.5
                  },
                  "categoria": {
                    "type": "string",
                    "example": "Papeleria"
                  },
                  "comprobante": {
                    "type": "string",
                    "example": "FAC-9001"
                  }
                },
                "required": [
                  "concepto",
                  "monto"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Gasto registrado"
          }
        }
      }
    },
    "/api/compras-suministro": {
      "get": {
        "tags": [
          "Caja - Cortes, Gastos y Compras"
        ],
        "summary": "Listar compras de suministro",
        "description": "Administrador o cajero.",
        "responses": {
          "200": {
            "description": "Listado de compras"
          }
        }
      },
      "post": {
        "tags": [
          "Caja - Cortes, Gastos y Compras"
        ],
        "summary": "Registrar compra de suministro",
        "description": "Administrador o cajero.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_caja": {
                    "type": "integer",
                    "example": 7
                  },
                  "proveedor": {
                    "type": "string",
                    "example": "Cafe del Valle S.A."
                  },
                  "total": {
                    "type": "number",
                    "example": 2400
                  },
                  "estado": {
                    "type": "string",
                    "enum": [
                      "pendiente",
                      "recibido"
                    ],
                    "example": "pendiente"
                  },
                  "factura": {
                    "type": "string",
                    "example": "FAC-SUM-2001"
                  },
                  "notas": {
                    "type": "string",
                    "example": "Pedido extraordinario de cafe en grano"
                  }
                },
                "required": [
                  "total"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Compra registrada"
          }
        }
      }
    },
...
  }
}